pyproject.toml
[tool.poetry]
name = "eth"
version = "0.1.0"
description = ""
authors = ["Jarvie8176 <3547456+Jarvie8176@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
web3 = "5.0.2"
greenlet = "^1.1.2"
pydantic = "^1.9.0"
loguru = "^0.6.0"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
pytz = "^2021.3"
pandas = "^1.4.1"
tronapi = "^3.1.6"

[tool.black]
line-length = 88


[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
coverage = { version = ">=6.3.2", extras = ["toml"] }
mypy = "^0.931"
pytest-dotenv = "^0.5.2"
flake8 = "^4.0.1"
flake8-tidy-imports = "^4.6.0"
flake8-pytest-style = "^1.6.0"
flake8-broken-line = "^0.4.0"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.8.0"
types-requests = "^2.27.11"
types-python-dateutil = "^2.8.9"
types-pytz = "^2021.3.5"
pytest-mypy = "^0.9.1"
pytest-mypy-plugins = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
concurrency = ['thread', 'greenlet']

[tool.coverage.report]
fail_under = 90
exclude_lines = [
  'pragma: no cover',
  'if TYPE_CHECKING:',
  'if __name__ == "__main__":',
  'raise NotImplementedError']

[tool.mypy]
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true
plugins = ['pydantic.mypy']
